pants goal clean-all-async
This does the same as clean-all, but does a quick rename of .pants.d and
then deletes the files in a spawned daemon process, rather than waiting
for the deletes. This is very helpful when building large scala projects,
which can generate tens of thousands of files.

(sapling split of 21bcf5c64384bd651cc8662371acbff31fd55621)