Fix pants invalidation bug.
Extra invalidation data had a fixed key for all targets in a non-flat
build, so the invalidation only applied to the first target, which
updated the key and made it appear to be unchanged for the subsequent
targets.

This change prefixes the key with the id of the set of targets.

(sapling split of 9c75e0a29de7694874f91230d0f455d60b6dac86)